imap/client: add FlagQuit()
If receiver waiting on a fetch command to complete and gets an
exit request, there's no way to interrupt the fetch. If using IMAP
directly, the call would just finish or error, unblocking execution.

When using the persistent client, it'll block the request indefinitely
until a connection is reestablished (by design). FlagQuit() is an
escape hatch to break out of the re-connection loop.

It will be used in a future patch.