receiver/flagcounter: make fields private and manage the channel internally
The channel is now closed instead of written to - this is so we don't
block.

Fixes a potential deadlock which happens if an IDLE has just finished,
i.e. Client.Idle() has returned but we haven't sent the
OperationIDLEFinish operation yet, and we receiver a quit request, we'll
try to write something to the "stop idle" channel. Since the IDLE's finished,
nothing's listening so we'll block.